02-tool-calls/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

google-api-python-client==2.190.0
google-auth-oauthlib==1.2.4
groq==1.0.0
python-dotenv==1.2.1
